requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
defusedxml==0.8.0rc2
Django==4.2.23
django-cors-headers==4.4.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==3.1.0
drf-yasg==1.21.10
google-auth==2.40.3
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
oauthlib==3.3.1
packaging==25.0
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.2.3
